added log function and fixed issues with capabilities
