--- tests/browserstack/allPlatforms.wdio.config.js
@@ -1,77 +1,0 @@
-const browserStackConfig = require('./browserstack.config');
-
-const date = new Date();
-const dateForDisplay = date.toDateString();
-const buildNameForDisplay = `${browserStackConfig.BUILD}: ${dateForDisplay}`;
-
-
-exports.config = {
-  user: browserStackConfig.BROWSERSTACK_USER,
-  key: browserStackConfig.BROWSERSTACK_KEY,
-  updateJob: false,
-  specs: [
-    './tests/browserstack/specs/test.js',
-  ],
-  exclude: [],
-  capabilities: [
-    // {
-    //   // Testing with a Windows Desktop Chrome browser
-    //   name: 'ballotMainTest-WindowsDesktopChromeBrowser',
-    //   build: buildNameForDisplay,
-    //   os: 'Windows',
-    //   os_version: '10',
-    //   browserName: 'Chrome',
-    //   browser_version: '72.0',
-    //   'browserstack.console': 'info',
-    //   'browserstack.debug': true,
-    //   'browserstack.geoLocation': 'US',
-    // },
-    {
-      // Testing with an iPhone Safari browser
-      name: 'ballotMainTest-iPhoneSafariBrowser',
-      build: buildNameForDisplay,
-      device: 'iPhone 6',
-      os_version: '11',
-      real_mobile: true,
-      browserName: 'Safari',
-      browser_version: '11',
-      isCordova: true,
-      foo: 'bar baz',
-      'browserstack.console': 'info',
-      'browserstack.debug': true,
-      'browserstack.geoLocation': 'US',
-    },
-    // {
-    //   // capabilities for a cordova android test
-    //   name: 'ballotMainTest-AndroidGalaxyTabS4',
-    //   build: buildNameForDisplay,
-    //   device: 'Samsung Galaxy Tab S4',
-    //   os_version: '8.1',
-    //   app: browserStackConfig.BROWSERSTACK_APK_URL,
-    //   'browserstack.console': 'info',
-    //   'browserstack.debug': true,
-    //   'browserstack.geoLocation': 'US',
-    // },
-    // {
-    //   // capabilities for a cordova iOS test
-    //   name: 'ballotMainTest-iOSiPhone6',
-    //   build: buildNameForDisplay,
-    //   device: 'iPhone 6',
-    //   os_version: '11',
-    //   app: browserStackConfig.BROWSERSTACK_IPA_URL,
-    //   'browserstack.debug': true,
-    //   'browserstack.geoLocation': 'US',
-    // },
-  ],
-  coloredLogs: true,
-  baseUrl: '',
-  waitforTimeout: 180000, // 3 minutes
-  connectionRetryTimeout: 90000,
-  connectionRetryCount: 3,
-
-  framework: 'mocha',
-  mochaOpts: {
-    ui: 'bdd',
-    timeout: 180000, // 3 minutes
-  },
-};
